@@ -4,11 +4,10 @@
     <router-view></router-view>
   </div>
 </template>
-
 <script>
   export default {}
 </script>
-<style lang="scss">
+<style lang="scss" rel="stylesheet/scss">
   .vue-notifyjs.notifications{
     .list-move {
       transition: transform 0.3s, opacity 0.4s;
@@ -16,7 +15,6 @@
     .list-item {
       display: inline-block;
       margin-right: 10px;
-
     }
     .list-enter-active {
       transition: transform 0.2s ease-in, opacity 0.4s ease-in;
@@ -24,11 +22,9 @@
     .list-leave-active {
       transition: transform 1s ease-out, opacity 0.4s ease-out;
     }
-
     .list-enter {
       opacity: 0;
       transform: scale(1.1);
-
     }
     .list-leave-to {
       opacity: 0;
