@@ -620,7 +620,7 @@
   </div>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from '../../UIComponents/Cards/Card.vue'
   export default {
     components: {
       Card
