@@ -45,7 +45,7 @@
   </div>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from '../UIComponents/Cards/Card.vue'
   export default {
     components: {
       Card
