@@ -4,7 +4,7 @@
       <div class="row">
         <div class="col-md-12">
           <card>
-            
+
           </card>
         </div>
       </div>
@@ -12,7 +12,7 @@
   </div>
 </template>
 <script>
-  import Card from 'src/components/UIComponents/Cards/Card.vue'
+  import Card from '../UIComponents/Cards/Card.vue'
   export default {
     components: {
       Card
