@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import App from './App.vue'
-import ErrorPage from 'src/components/GeneralViews/ErrorPage.vue'
+import ErrorPage from './components/GeneralViews/ErrorPage.vue'
 import BootstrapVue from 'bootstrap-vue'
 // apply global conf
 import config from './config.json'
